Only use source addresses in handlePacket
MulticastPacketReader uses an InetSocketAddress as a buffer that gets
filled for each incoming packet for the source address. Similarly to the
receive buffer, it should not be used outside of the handlePacket
method, as it will be modified on the handler thread when further
packets are received.

Fix the current usage in MdnsInterfaceAdvertiser/MdnsRecordRepository.
Fortunately due to a bug packets are never replied unicast, so today the
MdnsRecordRepository code using the source address never triggers, even
though it may be circulating an incorrect source address instance that
changes over time.

Test: atest
Change-Id: I4cadfa6e54d1b37d9a6b0f8ca0269193a229bc51

Former-commit-id: 7078ac59b9108979a2f598b77e8cfe8058ea7817